+++ components/cards/ContactMethodCard.tsx
@@ -18,7 +18,7 @@ const ContactMethodCard = ({ method }: ContactMethodCardProps) => {
         <p className="text-lg md:text-xl lg:text-2xl font-semibold text-amber-500">
           {method.name}
         </p>
-        <p className="text-sm md:text-base lg:text-lg">{method.description}</p>
+        <p className="text-sm md:text-base lg:text-lg">{method?.description}</p>
       </div>
     </Link>
   );