+++ components/sections/Contact.tsx
@@ -9,7 +9,7 @@ const Contact = () => {
         <SectionHeader sequence={["get in touch", 4000, "联络我", 4000]}>
           <span className="text-purple-500">contact</span> me
         </SectionHeader>
-        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-5 place-items-center">
+        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-5 place-items-center">
           {contactMethods.map((method, index) => (
             <ContactMethodCard method={method} key={index} />
           ))}